Activities corrected: code of new task moved from MainActivity to NewTaskActivity
