+++ app/src/main/res/layout/activity_main.xml
@@ -1,74 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".MainActivity"
-    android:background="@drawable/bg">
+    android:layout_height="wrap_content"
+    tools:context=".MainActivity">
 
-    <LinearLayout
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/recycler_view"
         android:layout_width="match_parent"
-        android:layout_height="72dp"
-        android:background="@color/white"
-        android:orientation="horizontal"
-        android:gravity="center">
+        android:layout_height="match_parent"
+        />
 
-        <EditText
-            android:id="@+id/etTaskName"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:hint="@string/et_task_name"
-            android:textColorHint="@color/black_new_item_hint"
-            android:textSize="16sp"
-            android:lineSpacingExtra="8sp"
-            android:layout_margin="16dp"
-            android:background="@null"
-            android:layout_weight="1"
-            />
-
-        <ImageButton
-            android:id="@+id/ibNewItem"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:src="@mipmap/ic_add_new_item"
-            android:layout_margin="16dp"
-            android:background="@color/white"
-            />
-
-    </LinearLayout>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:background="@android:color/darker_gray"/>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="48dp"
-        android:gravity="center_vertical">
-
-        <TextView
-            android:id="@+id/tvDot"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/tvDot"
-            android:textSize="30sp"
-            android:layout_marginLeft="16dp"
-            />
-
-        <TextView
-            android:id="@+id/tvPriority"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/tv_priority"
-            android:textSize="14sp"
-            android:lineSpacingExtra="10sp"
-            android:layout_marginLeft="8dp"
-            android:background="@color/white"
-            />
-
-    </LinearLayout>
-
-</LinearLayout>
+</FrameLayout>
